Move the PDF reading stack out of redact-pdf and into shared/js
Where every word on a page is, and which font drew it, is not a question
about redaction. redact-pdf answered it first because it was the first tool
that had to, and the answer has sat in tools/redact-pdf/src/ ever since:
base14.js, content.js, fonts.js, text.js, and the string codec at the top of
strings.js. The statement converter needs exactly that and nothing else from
the tool, so the modules move rather than get copied - test_duplicates.py
declares GROUPS empty and it is meant to stay that way.

Four files move whole. strings.js is split instead, because only decodeText
and encodeText are about reading a PDF: the rest of it sweeps every text
string in a document looking for words to remove, which is redaction and
belongs where it is. What stays re-exports the two so its callers do not have
to know the codec moved.

The reading stack leaves nothing behind that a visitor could see - the moved
modules throw no messages and call no phrase(), they return data and let the
caller say what happened - so the three prose sentences the ratchet counted
in them were comments. redact-pdf drops from 9 to 6 for that reason alone.

Verified after the move by building redact-pdf and opening it: main.js links,
no boot warning, and the example statement reads back as 3 pages and 237
placed glyphs.
